more output to console, new methods added for map functionality
size function added to retrieve size of map in elements. keyExists function added to check when removing a kv pair and also when retrieving what val is held at key. isEmpty function added to allow to check whether map is empty when removing elements. isFull function added to check if map is full, added this because only want to build new array if the map is full.